Relocate Message to components dir
